MDLSITE-2672 Avoid double "core_" prefix of subsystems
moodlecheck accepts both the list of plugins and subsystems
available in the site being run and, for external checks,
a .txt list of valid components from the CLI.

Recently the format of the list was changed to always include
complete component names (core_backup...) instead of simple
subsystem names (backup...).

With this change moodlecheck will accept both formats, adding
the 'core_' prefix only when needed.