MDLSITE-3940 control: prevent any PHP output when Docs are down
local_moodlecheck has a built-in mechanism/falbackk to use
a local file to extract the list of valid APIs within Moodle
when docs.moodle.org is unreachable,

And it works ok, but leads to the output of a PHP warning message
that ends breaking the expected (XML) output.

This simple fix just silences any output when that network operation
fails.