MDL-60786 Stop requiring @copyright and @license for individual classes
This made sense in the past where we typically had many classes in one
library file and it was useful to keep track of who holds copyright and
credit for each of the added class.

But today, with autoloading being the common and recommended practise,
we only have one class per file. So the information about the copyright
and license is typically duplicated - once in the file's phpdoc block
and then the same info in the class' block.