Add Github Actions to local_moodlecheck
Also, add the ability to travis runs to log-in to
DockerHub to avoid pulling rate limits.

With this patch, if both DOCKER_USER and DOCKER_TOKEN are defined
as environmental variables, they will be used to perform a login
to dockerhub, getting rid of pull limits.

Exactly the same than https://github.com/moodlehq/moodle-local_codechecker/pull/133
but applied to this repo.