Stop requiring the file phpdoc block for 1-artifact files
Now we detect all the classes, interfaces and traits in a file
and, if there is only one, then we stop requiring the file
phpdoc block.

This was agreed @ https://tracker.moodle.org/browse/MDLSITE-2804
and https://github.com/moodlehq/moodle-local_moodlecheck/issues/66
was created to implement it.

Note this doesn't address the whole agreement, but just the
detail about stop requiring the file phpdoc block, because it
is specially annoying and recurring in new files (that use to
be 1-artifact ones).