Ensure that only the final part of namespaces is returned
This tool has limited namespace handling capabilities and
we always have been getting, exclusively, the last part of
every namespace detected in function params.

This was the default behavior for PHP7 tokenizer and, in fact,
when reading the phpdocs for those params... we are removing
everything but the final part of the namespace too.

But PHP8 changed the default behavior and its tokenizer now
returns the complete namespace as unique token.

https://wiki.php.net/rfc/namespaced_names_as_token

So, to keep everything working as it was... when we detect that
some type of the function params contains namespaces, we are also
going to remove everything but the final part of the namespace.

That matches previous behaviour 100%.

If some day we want to go processing full namespaces (instead of
their final parts), that will be great, but not now.

Fixes #72.