Perform some php8 runs with 311 and master
Only real required change  is to set the new
max_input_vars=5000 requirement. Strictly speaking
only needed for php8, but it doesn't hurt with
older php versions and older moodle branches, so
setting it up always.

About the xmlrpc-beta extension, it's only needed
for php8 runs and only if the plugin uses xmlrpc
for something. Else, it can be omitted without
problem, as far as it's not a moodle requirement
but only a recommendation.

Finally, also bump version.php to current date
and make Moodle 3.5 (2018051700) required. This is
to make a new release in the plugins database.