Support interfaces and traits
Previously, only the class containing a function was retrieved and used in either rule. Now, the containing artifact is also looked up if it is an interface or a trait. Function objects contain a new 'owner' property for this. The old 'class' property remains, but can probably be removed unless third-party rules need to be supported.

To include the correct artifact type ('class', 'interface', or 'trait') in warnings, artifacts additional gain the property 'typestring' containing one of those three strings.

Fixes #96