Allow trailing comma in method parameters to work ok
Since PHP 8.0, trailing commas in method parameters are allowed, and
that is leading to an extra "phantom" parameter to be added to the
list of parameters detected (because of the way we manage tokens.

With this change, any empty returned token (without name and type)
will be ignored, because that's the case when parsing the params
and they have trailing commas.

Covered with tests using all the possible combinations of
method declarations (all them valid).