Remove all checks for define() (#141)
These checks never worked. The `get_defines()` method built a list of
definitions, added them to a `$defines` array, but return
`$this->defines` instead.

Furthermore there would be pandamonium if we actually fixed this because
we use constants in scripts all over Moodle with the same value, for
example:

```
define('CLI_SCRIPT', true);
```